WIP Introduce xcp-metrics-plugin-xen plugin for xcp-rrdd
xcp-metrics-plugin-xen is meant as a replacement for
- the metrics collection built into xcp-rrdd, which requires the
  latter to depend on libxenctrl
- the rrdp-pm plugin, which also fetches data through libxenctrl

Metrics are collected using the  Rust crate, structured in a
suitable way for upcoming OpenMetrics support, and communicated to
xcp-rrdd using its v2 protocol.

NOTE: at the time of this writing, we still rely on extra features we
had to add to the  crate, PR pending.

Signed-off-by: Teddy Astie <teddy.astie@outlook.fr>
Reviewed-by: Yann Dirson <yann.dirson@vates.fr>